backend/pyproject.toml
[project]
name = "encode-faces"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.*"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "azure-identity>=1.22.0",
    "azure-storage-blob>=12.25.1",
    "dlib>=19.24.8",
    "face-recognition>=1.3.0",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.2",
    "jwt>=1.3.1",
    "loguru>=0.7.3",
    "pgvector>=0.4.1",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-dotenv>=0.5.2",
    "setuptools>=79.0.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
